pyproject.toml
[tool.poetry]
name = "deepopg-off-the-shelf"
version = "0.1.0"
description = ""
authors = ["Harry Hsu <harry19930924@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.10"
setuptools = "59.5.0"

# this is painful to the eye but peaceful for your mind
# ... and only works for python3.9 + cuda 11.3
torch = { url = "https://download.pytorch.org/whl/cu113/torch-1.10.2%2Bcu113-cp39-cp39-linux_x86_64.whl", python = ">=3.9,<3.10", markers = "sys_platform == 'linux'" }
torchvision = { url = "https://download.pytorch.org/whl/cu113/torchvision-0.11.3%2Bcu113-cp39-cp39-linux_x86_64.whl", python = ">=3.9,<3.10", markers = "sys_platform == 'linux'" }
detectron2 = { url = "https://dl.fbaipublicfiles.com/detectron2/wheels/cu113/torch1.10/detectron2-0.6%2Bcu113-cp39-cp39-linux_x86_64.whl", python = ">=3.9,<3.10", markers = "sys_platform == 'linux'" }

# detectron2 deps
matplotlib = "^3.7.1"
numpy = "^1.24.2"
opencv-python = "^4.7.0.68"
pycocotools = "2.0.6"
scipy = "^1.10.1"
timm = "^0.6.13"
tqdm = "^4.65.0"

# extra deps
absl-py = "^1.4.0"
data-science-types = "^0.2.23"
ijson = "^3.2.0.post0"
imageio = "^2.25.1"
pandas = "^1.5.3"
pydantic = "^1.10.7"
rich = "^13.3.5"
types-requests = "^2.31.0.1"
pyomo = { extras = ["optional"], version = "^6.6.1" }

[tool.poetry.group.dev.dependencies]
mypy = "^1.2.0"
requests = "^2.31.0"

[tool.pyright]
include = ["app"]
exclude = ["**/__pycache__"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
